# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Sistema POS",
        "description": "Documentación de la API del Sistema POS",
        "contact": {
            "email": "admin@sistema.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/ajustes-inventario": {
            "get": {
                "tags": [
                    "Ajustes de Inventario"
                ],
                "summary": "Obtener lista de ajustes de inventario",
                "operationId": "99e036e9c1107b032163d643c8091c67",
                "parameters": [
                    {
                        "name": "tipo",
                        "in": "query",
                        "description": "Filtrar por tipo de ajuste (ENTRADA/SALIDA)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "ENTRADA",
                                "SALIDA"
                            ]
                        }
                    },
                    {
                        "name": "estado",
                        "in": "query",
                        "description": "Filtrar por estado del ajuste",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "PENDIENTE",
                                "COMPLETADO",
                                "ANULADO"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Cantidad de registros por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de ajustes de inventario",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "numero": {
                                                        "type": "string"
                                                    },
                                                    "tipo": {
                                                        "type": "string"
                                                    },
                                                    "motivo": {
                                                        "type": "string"
                                                    },
                                                    "estado": {
                                                        "type": "string"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "403": {
                        "description": "Prohibido - No tiene permiso para ver ajustes"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Ajustes de Inventario"
                ],
                "summary": "Crear un nuevo ajuste de inventario",
                "operationId": "80bbb65e617276e7524cea064cf5818a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "tipo",
                                    "motivo",
                                    "detalles"
                                ],
                                "properties": {
                                    "tipo": {
                                        "type": "string",
                                        "enum": [
                                            "ENTRADA",
                                            "SALIDA"
                                        ]
                                    },
                                    "motivo": {
                                        "description": "Motivo del ajuste",
                                        "type": "string"
                                    },
                                    "nota": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "detalles": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "producto_id",
                                                "unidad_medida_id",
                                                "cantidad",
                                                "costo"
                                            ],
                                            "properties": {
                                                "producto_id": {
                                                    "type": "integer"
                                                },
                                                "unidad_medida_id": {
                                                    "type": "integer"
                                                },
                                                "cantidad": {
                                                    "type": "number",
                                                    "format": "float"
                                                },
                                                "costo": {
                                                    "type": "number",
                                                    "format": "float"
                                                },
                                                "nota": {
                                                    "type": "string",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ajuste de inventario creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "numero": {
                                            "type": "string"
                                        },
                                        "tipo": {
                                            "type": "string"
                                        },
                                        "motivo": {
                                            "type": "string"
                                        },
                                        "estado": {
                                            "type": "string"
                                        },
                                        "detalles": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "producto": {
                                                        "type": "object"
                                                    },
                                                    "cantidad": {
                                                        "type": "number"
                                                    },
                                                    "costo": {
                                                        "type": "number"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "403": {
                        "description": "Prohibido - No tiene permiso para crear ajustes"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/ajustes-inventario/{id}": {
            "get": {
                "tags": [
                    "Ajustes de Inventario"
                ],
                "summary": "Obtener detalle de un ajuste de inventario",
                "operationId": "0b6665aa04dbc0c34866963785144ae7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle del ajuste de inventario",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "numero": {
                                            "type": "string"
                                        },
                                        "tipo": {
                                            "type": "string"
                                        },
                                        "motivo": {
                                            "type": "string"
                                        },
                                        "estado": {
                                            "type": "string"
                                        },
                                        "user": {
                                            "type": "object"
                                        },
                                        "detalles": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ajuste no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/ajustes-inventario/{id}/completar": {
            "post": {
                "tags": [
                    "Ajustes de Inventario"
                ],
                "summary": "Completar un ajuste de inventario",
                "operationId": "782c5d8a8c117cf9b7d6662a0beba673",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ajuste completado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error al completar el ajuste",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ajuste no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/ajustes-inventario/{id}/anular": {
            "post": {
                "tags": [
                    "Ajustes de Inventario"
                ],
                "summary": "Anular un ajuste de inventario",
                "operationId": "7e70122c145d794bef0e1351c52a0025",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ajuste anulado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error al anular el ajuste",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ajuste no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}